% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_operations.R
\name{batch_check_layer_availability}
\alias{batch_check_layer_availability}
\title{Check the availability of multiple image layers in a specified registry and repository}
\usage{
batch_check_layer_availability(repositoryName, layerDigests,
  registryId = NULL)
}
\arguments{
\item{repositoryName}{The name of the repository that is associated with the image layers to check.}

\item{layerDigests}{The digests of the image layers to check.}

\item{registryId}{The AWS account ID associated with the registry that contains the image layers to check. If you do not specify a registry, the default registry is assumed.}
}
\description{
Check the availability of multiple image layers in a specified registry and repository.
}
\details{
This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers for pulling and pushing images. In most cases, you should use the \code{docker} CLI to pull, tag, and push images.
}
\examples{

}
