% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_operations.R
\name{complete_layer_upload}
\alias{complete_layer_upload}
\title{Informs Amazon ECR that the image layer upload has completed for a specified registry, repository name, and upload ID}
\usage{
complete_layer_upload(repositoryName, uploadId, layerDigests,
  registryId = NULL)
}
\arguments{
\item{repositoryName}{The name of the repository to associate with the image layer.}

\item{uploadId}{The upload ID from a previous InitiateLayerUpload operation to associate with the image layer.}

\item{layerDigests}{The \code{sha256} digest of the image layer.}

\item{registryId}{The AWS account ID associated with the registry to which to upload layers. If you do not specify a registry, the default registry is assumed.}
}
\description{
Informs Amazon ECR that the image layer upload has completed for a specified registry, repository name, and upload ID. You can optionally provide a \code{sha256} digest of the image layer for data validation purposes.
}
\details{
This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers for pulling and pushing images. In most cases, you should use the \code{docker} CLI to pull, tag, and push images.
}
\examples{

}
