% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_operations.R
\name{create_repository}
\alias{create_repository}
\title{Creates an image repository}
\usage{
create_repository(repositoryName, tags = NULL)
}
\arguments{
\item{repositoryName}{[required] The name to use for the repository. The repository name may be specified on its own (such as \code{nginx-web-app}) or it can be prepended with a namespace to group the repository into a category (such as \code{project-a/nginx-web-app}).}

\item{tags}{}
}
\description{
Creates an image repository.
}
\section{Accepted Parameters}{
\preformatted{create_repository(
  repositoryName = "string",
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\examples{
# This example creates a repository called nginx-web-app inside the
# project-a namespace in the default registry for an account.
\donttest{create_repository(
  repositoryName = "project-a/nginx-web-app"
)}

}
