% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_operations.R
\name{get_lifecycle_policy}
\alias{get_lifecycle_policy}
\title{Retrieves the specified lifecycle policy}
\usage{
get_lifecycle_policy(registryId = NULL, repositoryName)
}
\arguments{
\item{registryId}{The AWS account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.}

\item{repositoryName}{[required] The name of the repository.}
}
\description{
Retrieves the specified lifecycle policy.
}
\section{Accepted Parameters}{
\preformatted{get_lifecycle_policy(
  registryId = "string",
  repositoryName = "string"
)
}
}

