% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_operations.R
\name{get_repository_policy}
\alias{get_repository_policy}
\title{Retrieves the repository policy for a specified repository}
\usage{
get_repository_policy(repositoryName, registryId = NULL)
}
\arguments{
\item{repositoryName}{The name of the repository with the policy to retrieve.}

\item{registryId}{The AWS account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.}
}
\description{
Retrieves the repository policy for a specified repository.
}
\examples{
# This example obtains the repository policy for the repository named
# ubuntu.
get_repository_policy(
  repositoryName = "ubuntu"
)

}
