% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{List the tags for an Amazon ECR resource}
\usage{
list_tags_for_resource(
  resourceArn = "string"
)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource for which to list the tags. Currently, the only supported resource is an Amazon ECR repository.}
}
\description{
List the tags for an Amazon ECR resource.
}
\examples{

}
