% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_package.R
\docType{package}
\name{paws.ecr-package}
\alias{paws.ecr}
\alias{paws.ecr-package}
\title{paws.ecr: Amazon EC2 Container Registry}
\description{
Amazon Elastic Container Registry (Amazon ECR) is a
managed Docker registry service. Customers can use the familiar Docker
CLI to push, pull, and manage images. Amazon ECR provides a secure,
scalable, and reliable registry. Amazon ECR supports private Docker
repositories with resource-based permissions using IAM so that
specific users or Amazon EC2 instances can access repositories and
images. Developers can use the Docker CLI to author and manage images.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
