% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_operations.R
\name{put_image}
\alias{put_image}
\title{Creates or updates the image manifest and tags associated with an image}
\usage{
put_image(
  registryId = "string",
  repositoryName = "string",
  imageManifest = "string",
  imageTag = "string"
)
}
\arguments{
\item{registryId}{The AWS account ID associated with the registry that contains the repository in which to put the image. If you do not specify a registry, the default registry is assumed.}

\item{repositoryName}{[required] The name of the repository in which to put the image.}

\item{imageManifest}{[required] The image manifest corresponding to the image to be uploaded.}

\item{imageTag}{The tag to associate with the image. This parameter is required for images that use the Docker Image Manifest V2 Schema 2 or OCI formats.}
}
\description{
Creates or updates the image manifest and tags associated with an image.
}
\details{
This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers for pulling and pushing images. In most cases, you should use the \code{docker} CLI to pull, tag, and push images.
}
\examples{

}
