% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_operations.R
\name{put_lifecycle_policy}
\alias{put_lifecycle_policy}
\title{Creates or updates a lifecycle policy}
\usage{
put_lifecycle_policy(repositoryName, lifecyclePolicyText, registryId = NULL)
}
\arguments{
\item{repositoryName}{The name of the repository to receive the policy.}

\item{lifecyclePolicyText}{The JSON repository policy text to apply to the repository.}

\item{registryId}{The AWS account ID associated with the registry that contains the repository. If you do  not specify a registry, the default registry is assumed.}
}
\description{
Creates or updates a lifecycle policy. For information about lifecycle policy syntax, see \href{http://docs.aws.amazon.com/AmazonECR/latest/userguide/LifecyclePolicies.html}{Lifecycle Policy Template}.
}
\examples{

}
