% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Deletes specified tags from a resource}
\usage{
untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource from which to remove tags. Currently, the only supported resource is an Amazon ECR repository.}

\item{tagKeys}{[required] The keys of the tags to be removed.}
}
\description{
Deletes specified tags from a resource.
}
\examples{

}
