% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_operations.R
\name{delete_lifecycle_policy}
\alias{delete_lifecycle_policy}
\title{Deletes the specified lifecycle policy}
\usage{
delete_lifecycle_policy(
  registryId = "string",
  repositoryName = "string"
)
}
\arguments{
\item{registryId}{The AWS account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.}

\item{repositoryName}{[required] The name of the repository.}
}
\description{
Deletes the specified lifecycle policy.
}
\examples{

}
