% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecr_operations.R
\name{get_repository_policy}
\alias{get_repository_policy}
\title{Retrieves the repository policy for a specified repository}
\usage{
get_repository_policy(registryId = NULL, repositoryName)
}
\arguments{
\item{registryId}{The AWS account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.}

\item{repositoryName}{[required] The name of the repository with the policy to retrieve.}
}
\description{
Retrieves the repository policy for a specified repository.
}
\section{Accepted Parameters}{
\preformatted{get_repository_policy(
  registryId = "string",
  repositoryName = "string"
)
}
}

\examples{
# This example obtains the repository policy for the repository named
# ubuntu.
\donttest{get_repository_policy(
  repositoryName = "ubuntu"
)}

}
